refactor: Renamed all packages to singular After doing more research & comparing our project structure with more professionals git, I see that we should keep our packages name singular
