+++ src/main/java/edu/lorsenmarek/backend/repository/SerieRepository.java
@@ -1,5 +1,5 @@
-package edu.lorsenmarek.backend.repositories;
-import edu.lorsenmarek.backend.models.Serie;
+package edu.lorsenmarek.backend.repository;
+import edu.lorsenmarek.backend.model.Serie;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
